Prepare for version 1.2.
Fix the bug of empty conflict group in conflict handling.
Fix that packages with the prefix lib32- or lib64- cannot be selected through the default configuration file.
Fix the bug that the package could not be removed correctly when removing.
Fix the bug that the config file exits abnormally when the file name is empty.
Fix the bug of using the config file to install the installed package.
Fix the bug that cannot be packaged.

Signed-off-by: Zang Ruochen <zangrc.fnst@cn.fujitsu.com>
Signed-off-by: Zheng Ruoqin <zhengrq.fnst@cn.fujitsu.com>